backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
sqlalchemy==2.0.35
pydantic-settings==2.4.0
pymysql==1.1.0
cryptography==43.0.1
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
bcrypt==5.0.0

# Phase 2: AI models & data pipeline
transformers>=4.40.0
sentence-transformers>=2.7.0
torch>=2.9.0
gitpython>=3.1.40
chromadb>=1.5.0
elasticsearch>=8.13.0,<9.0.0
numpy>=1.26.0

# Phase 3: LangGraph reasoning agent
langgraph>=0.1.0
langchain-core>=0.2.0

# Phase 4: API authentication & multipart support
python-multipart>=0.0.9
email-validator>=2.1.0
httpx>=0.27.0
